+++ main_test.go
@@ -13,7 +13,7 @@ import (
 	"testing"
 )
 
-func createConfig(ec2 bool, ecs bool, fis bool, lambda bool, rds bool, zone bool) config.Specification {
+func createConfig(ec2 bool, ecs bool, elb bool, fis bool, lambda bool, rds bool, zone bool) config.Specification {
 	return config.Specification{
 		EnrichEc2DataForTargetTypes: []string{"com.steadybit.extension_aws.test"},
 		DiscoveryDisabledEc2:        ec2,
@@ -21,6 +21,8 @@ func createConfig(ec2 bool, ecs bool, fis bool, lambda bool, rds bool, zone bool
 		DiscoveryDisabledEcs:        ecs,
 		DiscoveryIntervalEcsTask:    10,
 		DiscoveryIntervalEcsService: 10,
+		DiscoveryDisabledElb:        elb,
+		DiscoveryIntervalElbAlb:     10,
 		DiscoveryDisabledFis:        fis,
 		DiscoveryIntervalFis:        10,
 		DiscoveryDisabledLambda:     lambda,
@@ -40,7 +42,7 @@ func Test_getExtensionList(t *testing.T) {
 	}{
 		{
 			name:   "disabled all but ec2",
-			config: createConfig(false, true, true, true, true, true),
+			config: createConfig(false, true, true, true, true, true, true),
 			wantedRoutes: []string{
 				"/com.steadybit.extension_aws.ec2_instance.state",
 				"/com.steadybit.extension_aws.ec2-instance/discovery",
@@ -53,7 +55,7 @@ func Test_getExtensionList(t *testing.T) {
 		},
 		{
 			name:   "disabled all but ecs",
-			config: createConfig(true, false, true, true, true, true),
+			config: createConfig(true, false, true, true, true, true, true),
 			wantedRoutes: []string{
 				"/com.steadybit.extension_aws.ecs-task.stop",
 				"/com.steadybit.extension_aws.ecs-service.scale",
@@ -70,8 +72,17 @@ func Test_getExtensionList(t *testing.T) {
 			},
 		},
 		{
+			name:   "disabled all but elb",
+			config: createConfig(true, true, false, true, true, true, true),
+			wantedRoutes: []string{
+				"/com.steadybit.extension_aws.alb/discovery",
+				"/com.steadybit.extension_aws.alb/discovery/target-description",
+				"/discovery/attributes",
+			},
+		},
+		{
 			name:   "disabled all but fis",
-			config: createConfig(true, true, false, true, true, true),
+			config: createConfig(true, true, true, false, true, true, true),
 			wantedRoutes: []string{
 				"/com.steadybit.extension_aws.fis.start_experiment",
 				"/com.steadybit.extension_aws.fis-experiment-template/discovery",
@@ -81,7 +92,7 @@ func Test_getExtensionList(t *testing.T) {
 		},
 		{
 			name:   "disabled all but lambda",
-			config: createConfig(true, true, true, false, true, true),
+			config: createConfig(true, true, true, true, false, true, true),
 			wantedRoutes: []string{
 				"/com.steadybit.extension_aws.lambda.denylist",
 				"/com.steadybit.extension_aws.lambda.diskspace",
@@ -95,7 +106,7 @@ func Test_getExtensionList(t *testing.T) {
 		},
 		{
 			name:   "disabled all but rds",
-			config: createConfig(true, true, true, true, false, true),
+			config: createConfig(true, true, true, true, true, false, true),
 			wantedRoutes: []string{
 				"/com.steadybit.extension_aws.rds.cluster.failover",
 				"/com.steadybit.extension_aws.rds.instance.reboot",
@@ -109,7 +120,7 @@ func Test_getExtensionList(t *testing.T) {
 		},
 		{
 			name:   "disabled all but zone",
-			config: createConfig(true, true, true, true, true, false),
+			config: createConfig(true, true, true, true, true, true, false),
 			wantedRoutes: []string{
 				"/com.steadybit.extension_aws.az.blackhole",
 				"/com.steadybit.extension_aws.zone/discovery",